Refactor : Validation 추가된 테스트코드 수정
