+++ orury-client/src/test/java/org/orury/client/crew/application/CrewServiceImplTest.java
@@ -13,7 +13,6 @@ import org.orury.client.global.image.ImageAsyncStore;
 import org.orury.common.error.code.CrewErrorCode;
 import org.orury.common.error.exception.BusinessException;
 import org.orury.common.util.S3Folder;
-import org.orury.domain.DomainFixtureFactory;
 import org.orury.domain.crew.domain.*;
 import org.orury.domain.crew.domain.dto.CrewDto;
 import org.orury.domain.crew.domain.dto.CrewGender;
@@ -163,34 +162,6 @@ class CrewServiceImplTest {
                 .addCrewMember(anyLong(), anyLong());
     }
 
-    @DisplayName("[createCrew] 크루 참여/신청 횟수가 5 이상이면, MaximumParticipation 예외를 발생시킨다.")
-    @Test
-    void when_OverMaximumParticipation_Then_MaximumParticipationException() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        MultipartFile file = mock(MultipartFile.class);
-        int participatingCrewCount = 2;
-        int applyingCrewCount = 3;
-        given(crewMemberReader.countByUserId(anyLong()))
-                .willReturn(participatingCrewCount);
-        given(crewApplicationReader.countByUserId(anyLong()))
-                .willReturn(applyingCrewCount);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.createCrew(crewDto, file));
-
-        assertEquals(CrewErrorCode.MAXIMUM_PARTICIPATION.getMessage(), exception.getMessage());
-        then(crewMemberReader).should(only())
-                .countByUserId(anyLong());
-        then(crewApplicationReader).should(only())
-                .countByUserId(anyLong());
-        then(imageAsyncStore).shouldHaveNoInteractions();
-        then(crewStore).shouldHaveNoInteractions();
-        then(crewTagStore).shouldHaveNoInteractions();
-        then(crewMemberStore).shouldHaveNoInteractions();
-    }
-
     @DisplayName("[getCrewDtosByRank] 페이지와 랭킹에 따른 크루 목록을 가져온다.")
     @Test
     void should_GetCrewDtosByRank() {
@@ -291,6 +262,24 @@ class CrewServiceImplTest {
     @DisplayName("[updateCrewInfo] 크루 정보를 업데이트한다.")
     @Test
     void should_UpdateCrewInfo() {
+        // given
+        CrewDto oldCrew = createCrewDto().build().get();
+        CrewDto newCrew = createCrewDto().build().get();
+        Long userId = oldCrew.userDto().id();
+        Crew crew = createCrew().build().get();
+        given(crewStore.save(any()))
+                .willReturn(crew);
+
+        // when
+        crewService.updateCrewInfo(oldCrew, newCrew, userId);
+
+        // then
+        then(crewUpdatePolicy).should(only())
+                .validateUpdateCrewInfo(any(), any(), anyLong());
+        then(crewStore).should(only())
+                .save(any());
+        then(crewTagStore).should(only())
+                .updateTags(anyList(), anyList(), any());
     }
 
     @DisplayName("[updateCrewImage] 크루 이미지를 업데이트한다.")
@@ -316,24 +305,6 @@ class CrewServiceImplTest {
                 .delete(any(), anyString());
     }
 
-    @DisplayName("[updateCrewImage] 크루장이 아닌 경우, Forbidden 예외를 발생시킨다.")
-    @Test
-    void when_NotCrewCreator_Then_ForbiddenException1() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        MultipartFile file = mock(MultipartFile.class);
-        Long userId = 2134L;
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.updateCrewImage(crewDto, file, userId));
-
-        assertEquals(CrewErrorCode.FORBIDDEN.getMessage(), exception.getMessage());
-        then(imageAsyncStore).shouldHaveNoInteractions();
-        then(crewStore).shouldHaveNoInteractions();
-        then(imageStore).shouldHaveNoInteractions();
-    }
-
     @DisplayName("[deleteCrew] 크루를 삭제한다.")
     @Test
     void should_DeleteCrew() {
@@ -349,22 +320,6 @@ class CrewServiceImplTest {
                 .delete(any());
         then(imageStore).should(only())
                 .delete(any(), anyString());
-    }
-
-    @DisplayName("[deleteCrew] 크루장이 아닌 경우, Forbidden 예외를 발생시킨다.")
-    @Test
-    void when_NotCrewCreator_Then_ForbiddenException2() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        Long userId = 284L;
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.deleteCrew(crewDto, userId));
-
-        assertEquals(CrewErrorCode.FORBIDDEN.getMessage(), exception.getMessage());
-        then(crewStore).shouldHaveNoInteractions();
-        then(imageStore).shouldHaveNoInteractions();
     }
 
     @DisplayName("[applyCrew] 크루에 가입신청을 한다.")
@@ -383,212 +338,15 @@ class CrewServiceImplTest {
                 .birthday(LocalDate.now().minusYears(20)).build().get();
 
         String answer = "가입신청 답변";
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userDto.id()))
-                .willReturn(false);
-        given(crewMemberReader.countByUserId(userDto.id()))
-                .willReturn(2);
-        given(crewApplicationReader.countByUserId(userDto.id()))
-                .willReturn(2);
 
         // when
         crewService.applyCrew(crewDto, userDto, answer);
 
         // then
-        then(crewMemberReader).should(times(1))
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewMemberReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewApplicationReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewMemberStore).shouldHaveNoInteractions();
+        then(crewApplicationPolicy).should(only())
+                .validateApplyCrew(crewDto, userDto, answer);
         then(crewApplicationStore).should(only())
                 .save(any(), any(), anyString());
-    }
-
-    @DisplayName("[applyCrew] 크루원이 이미 존재하는 경우, AlreadyMember 예외를 발생시킨다.")
-    @Test
-    void when_AlreadyMember_Then_AlreadyMemberException() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        UserDto userDto = createUserDto().build().get();
-        String answer = "가입신청 답변";
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userDto.id()))
-                .willReturn(true);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.applyCrew(crewDto, userDto, answer));
-
-        assertEquals(CrewErrorCode.ALREADY_MEMBER.getMessage(), exception.getMessage());
-        then(crewMemberReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewMemberReader).should(never())
-                .countByUserId(anyLong());
-        then(crewApplicationReader).shouldHaveNoInteractions();
-        then(crewMemberStore).shouldHaveNoInteractions();
-        then(crewApplicationStore).shouldHaveNoInteractions();
-    }
-
-    @DisplayName("[applyCrew] 크루 참여/신청 횟수가 5 이상이면, MaximumParticipation 예외를 발생시킨다.")
-    @Test
-    void when_OverMaximumParticipation_Then_MaximumParticipationException1() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        UserDto userDto = createUserDto().build().get();
-        String answer = "가입신청 답변";
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userDto.id()))
-                .willReturn(false);
-        given(crewMemberReader.countByUserId(userDto.id()))
-                .willReturn(2);
-        given(crewApplicationReader.countByUserId(userDto.id()))
-                .willReturn(3);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.applyCrew(crewDto, userDto, answer));
-
-        assertEquals(CrewErrorCode.MAXIMUM_PARTICIPATION.getMessage(), exception.getMessage());
-        then(crewMemberReader).should(times(1))
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewMemberReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewApplicationReader).should(only())
-                .countByUserId(anyLong());
-        then(crewMemberStore).shouldHaveNoInteractions();
-        then(crewApplicationStore).shouldHaveNoInteractions();
-    }
-
-    @DisplayName("[applyCrew] 지원자의 연령이 크루 연령기준에 부합하지 않으면, AgeForbidden 예외를 발생시킨다.")
-    @Test
-    void when_AgeForbidden_Then_AgeForbiddenException() {
-        // given
-        CrewDto crewDto = createCrewDto()
-                .id(23L)
-                .minAge(15)
-                .maxAge(30).build().get();
-        UserDto userDto = createUserDto()
-                .birthday(LocalDate.now().minusYears(14)).build().get();
-        String answer = "가입신청 답변";
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userDto.id()))
-                .willReturn(false);
-        given(crewMemberReader.countByUserId(userDto.id()))
-                .willReturn(2);
-        given(crewApplicationReader.countByUserId(userDto.id()))
-                .willReturn(2);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.applyCrew(crewDto, userDto, answer));
-
-        assertEquals(CrewErrorCode.AGE_FORBIDDEN.getMessage(), exception.getMessage());
-        then(crewMemberReader).should(times(1))
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewMemberReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewApplicationReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewMemberStore).shouldHaveNoInteractions();
-        then(crewApplicationStore).shouldHaveNoInteractions();
-    }
-
-    @DisplayName("[applyCrew] 크루원의 성별이 크루 성별기준에 부합하지 않으면, GenderForbidden 예외를 발생시킨다.")
-    @Test
-    void when_GenderForbidden_Then_GenderForbiddenException() {
-        // given
-        CrewDto crewDto = createCrewDto()
-                .gender(CrewGender.FEMALE).build().get();
-        UserDto userDto = createUserDto()
-                .gender(NumberConstants.MALE).build().get();
-        String answer = "가입신청 답변";
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userDto.id()))
-                .willReturn(false);
-        given(crewMemberReader.countByUserId(userDto.id()))
-                .willReturn(2);
-        given(crewApplicationReader.countByUserId(userDto.id()))
-                .willReturn(2);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.applyCrew(crewDto, userDto, answer));
-
-        assertEquals(CrewErrorCode.GENDER_FORBIDDEN.getMessage(), exception.getMessage());
-        then(crewMemberReader).should(times(1))
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewMemberReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewApplicationReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewMemberStore).shouldHaveNoInteractions();
-        then(crewApplicationStore).shouldHaveNoInteractions();
-    }
-
-    @DisplayName("[applyCrew] 지원하는 크루가 즉시 가입이고 신청자가 가입조건을 만족하는 경우, 신청자를 바로 가입시킨다.")
-    @Test
-    void when_PermissionNotRequiredCrew_Then_ImmediateJoin() {
-        // given
-        CrewDto crewDto = createCrewDto()
-                .minAge(25)
-                .maxAge(30)
-                .gender(CrewGender.FEMALE)
-                .permissionRequired(false)
-                .answerRequired(false).build().get();
-        UserDto userDto = createUserDto()
-                .birthday(LocalDate.now().minusYears(26))
-                .gender(NumberConstants.FEMALE).build().get();
-
-        String answer = "가입신청 답변";
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userDto.id()))
-                .willReturn(false);
-        given(crewMemberReader.countByUserId(userDto.id()))
-                .willReturn(2);
-        given(crewApplicationReader.countByUserId(userDto.id()))
-                .willReturn(2);
-
-        // when
-        crewService.applyCrew(crewDto, userDto, answer);
-
-        // then
-        then(crewMemberReader).should(times(1))
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewMemberReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewApplicationReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewMemberStore).should(times(1))
-                .addCrewMember(anyLong(), anyLong());
-        then(crewApplicationStore).shouldHaveNoInteractions();
-    }
-
-    @DisplayName("[applyCrew] 크루원의 가입신청 답변이 필수한데 답변이 없는 경우, EmptyAnswer 예외를 발생시킨다.")
-    @Test
-    void when_AnswerRequiredButNoAnswer_Then_AnswerRequiredException() {
-        // given
-        CrewDto crewDto = createCrewDto()
-                .gender(CrewGender.ANY)
-                .permissionRequired(true)
-                .answerRequired(true).build().get();
-        UserDto userDto = createUserDto().build().get();
-        String answer = "";
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userDto.id()))
-                .willReturn(false);
-        given(crewMemberReader.countByUserId(userDto.id()))
-                .willReturn(2);
-        given(crewApplicationReader.countByUserId(userDto.id()))
-                .willReturn(2);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.applyCrew(crewDto, userDto, answer));
-
-        assertEquals(CrewErrorCode.EMPTY_ANSWER.getMessage(), exception.getMessage());
-        then(crewMemberReader).should(times(1))
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewMemberReader).should(times(1))
-                .countByUserId(anyLong());
-        then(crewApplicationReader).should(only())
-                .countByUserId(anyLong());
-        then(crewMemberStore).shouldHaveNoInteractions();
-        then(crewApplicationStore).shouldHaveNoInteractions();
     }
 
     @DisplayName("[withdrawApplication] 크루 가입신청을 취소한다.")
@@ -597,36 +355,15 @@ class CrewServiceImplTest {
         // given
         Long userId = 1L;
         CrewDto crewDto = createCrewDto().build().get();
-        given(crewApplicationReader.existsByCrewIdAndUserId(crewDto.id(), userId))
-                .willReturn(true);
 
         // when
         crewService.withdrawApplication(crewDto, userId);
 
         // then
-        then(crewApplicationReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
+        then(crewApplicationPolicy).should(only())
+                .validateApplication(anyLong(), anyLong());
         then(crewApplicationStore).should(only())
                 .delete(any(), any());
-    }
-
-    @DisplayName("[withdrawApplication] 존재하지 않는 신청의 경우, NotFoundApplication 예외를 발생시킨다.")
-    @Test
-    void when_NotExistingApplication_Then_NotFoundApplicationException1() {
-        // given
-        Long userId = 1L;
-        CrewDto crewDto = createCrewDto().build().get();
-        given(crewApplicationReader.existsByCrewIdAndUserId(crewDto.id(), userId))
-                .willReturn(false);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.withdrawApplication(crewDto, userId));
-
-        assertEquals(CrewErrorCode.NOT_FOUND_APPLICATION.getMessage(), exception.getMessage());
-        then(crewApplicationReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewApplicationStore).shouldHaveNoInteractions();
     }
 
     @DisplayName("[approveApplication] 크루 가입신청을 승인한다.")
@@ -635,54 +372,19 @@ class CrewServiceImplTest {
         // given
         CrewDto crewDto = createCrewDto().build().get();
         Long applicantId = 2L;
-        given(crewApplicationReader.existsByCrewIdAndUserId(crewDto.id(), applicantId))
-                .willReturn(true);
 
         // when
         crewService.approveApplication(crewDto, applicantId, crewDto.userDto().id());
 
         // then
-        then(crewApplicationReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
+        then(crewPolicy).should(only())
+                .validateCrewCreator(anyLong(), anyLong());
+        then(crewApplicationPolicy).should(only())
+                .validateApplication(anyLong(), anyLong());
         then(crewApplicationStore).should(only())
                 .approve(anyLong(), anyLong());
     }
 
-    @DisplayName("[approveApplication] 크루장이 아닌 경우, Forbidden 예외를 발생시킨다.")
-    @Test
-    void when_NotCrewCreator_Then_ForbiddenException3() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        Long applicantId = 2L;
-        Long invalidUserId = 3L;
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.approveApplication(crewDto, applicantId, invalidUserId));
-
-        assertEquals(CrewErrorCode.FORBIDDEN.getMessage(), exception.getMessage());
-        then(crewApplicationReader).shouldHaveNoInteractions();
-        then(crewApplicationStore).shouldHaveNoInteractions();
-    }
-
-    @DisplayName("[approveApplication] 존재하지 않는 신청의 경우, NotFoundApplication 예외를 발생시킨다.")
-    @Test
-    void when_NotExistingApplication_Then_NotFoundApplicationException2() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        Long applicantId = 2L;
-        given(crewApplicationReader.existsByCrewIdAndUserId(crewDto.id(), applicantId))
-                .willReturn(false);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.approveApplication(crewDto, applicantId, crewDto.userDto().id()));
-
-        assertEquals(CrewErrorCode.NOT_FOUND_APPLICATION.getMessage(), exception.getMessage());
-        then(crewApplicationReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewApplicationStore).shouldHaveNoInteractions();
-    }
 
     @DisplayName("[disapproveApplication] 크루 가입신청을 거절한다.")
     @Test
@@ -690,53 +392,17 @@ class CrewServiceImplTest {
         // given
         CrewDto crewDto = createCrewDto().build().get();
         Long applicantId = 2L;
-        given(crewApplicationReader.existsByCrewIdAndUserId(crewDto.id(), applicantId))
-                .willReturn(true);
 
         // when
         crewService.disapproveApplication(crewDto, applicantId, crewDto.userDto().id());
 
         // then
-        then(crewApplicationReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
+        then(crewPolicy).should(only())
+                .validateCrewCreator(anyLong(), anyLong());
+        then(crewApplicationPolicy).should(only())
+                .validateApplication(anyLong(), anyLong());
         then(crewApplicationStore).should(only())
                 .delete(anyLong(), anyLong());
-    }
-
-    @DisplayName("[disapproveApplication] 크루장이 아닌 경우, Forbidden 예외를 발생시킨다.")
-    @Test
-    void when_NotCrewCreator_Then_ForbiddenException4() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        Long applicantId = 2L;
-        Long invalidUserId = 3L;
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.disapproveApplication(crewDto, applicantId, invalidUserId));
-
-        assertEquals(CrewErrorCode.FORBIDDEN.getMessage(), exception.getMessage());
-        then(crewApplicationReader).shouldHaveNoInteractions();
-        then(crewApplicationStore).shouldHaveNoInteractions();
-    }
-
-    @DisplayName("[disapproveApplication] 존재하지 않는 신청의 경우, NotFoundApplication 예외를 발생시킨다.")
-    @Test
-    void when_NotExistingApplication_Then_NotFoundApplicationException3() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        Long applicantId = 2L;
-        given(crewApplicationReader.existsByCrewIdAndUserId(crewDto.id(), applicantId))
-                .willReturn(false);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.disapproveApplication(crewDto, applicantId, crewDto.userDto().id()));
-
-        assertEquals(CrewErrorCode.NOT_FOUND_APPLICATION.getMessage(), exception.getMessage());
-        then(crewApplicationReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(crewApplicationStore).shouldHaveNoInteractions();
     }
 
     @DisplayName("[leaveCrew] 크루를 탈퇴한다.")
@@ -745,15 +411,13 @@ class CrewServiceImplTest {
         // given
         CrewDto crewDto = createCrewDto().build().get();
         Long userId = 11L;
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userId))
-                .willReturn(true);
 
         // when
         crewService.leaveCrew(crewDto, userId);
 
         // then
-        then(crewMemberReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
+        then(crewPolicy).should(only())
+                .validateCrewMember(anyLong(), anyLong());
         then(meetingMemberStore).should(only())
                 .removeAllByUserIdAndCrewId(anyLong(), anyLong());
         then(meetingStore).should(only())
@@ -780,27 +444,6 @@ class CrewServiceImplTest {
         then(crewMemberStore).shouldHaveNoInteractions();
     }
 
-    @DisplayName("[leaveCrew] 존재하지 않는 크루원의 경우, NotCrewMember 예외를 발생시킨다.")
-    @Test
-    void when_NotExistingCrewMember_Then_NotFoundCrewMemberException() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        Long userId = 11L;
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), userId))
-                .willReturn(false);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.leaveCrew(crewDto, userId));
-
-        assertEquals(CrewErrorCode.NOT_CREW_MEMBER.getMessage(), exception.getMessage());
-        then(crewMemberReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
-        then(meetingMemberStore).shouldHaveNoInteractions();
-        then(meetingStore).shouldHaveNoInteractions();
-        then(crewMemberStore).shouldHaveNoInteractions();
-    }
-
     @DisplayName("크루원을 추방한다.")
     @Test
     void should_ExpelMember() {
@@ -812,37 +455,16 @@ class CrewServiceImplTest {
         crewService.expelMember(crewDto, memberId, crewDto.userDto().id());
 
         // then
-        then(crewPolicy).should(only())
+        then(crewPolicy).should(times(1))
                 .validateCrewCreator(anyLong(), anyLong());
-        then(crewMemberReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
+        then(crewPolicy).should(times(1))
+                .validateCrewMember(anyLong(), anyLong());
         then(meetingMemberStore).should(only())
                 .removeAllByUserIdAndCrewId(anyLong(), anyLong());
         then(meetingStore).should(only())
                 .deleteAllByUserIdAndCrewId(anyLong(), anyLong());
         then(crewMemberStore).should(only())
                 .subtractCrewMember(anyLong(), anyLong());
-    }
-
-    @DisplayName("크루원을 추방할 때, 크루장이 아닌 경우, Forbidden 예외를 발생시킨다.")
-    @Test
-    void when_NotCrewCreator_Then_ForbiddenException5() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        Long memberId = 11L;
-        Long invalidUserId = 12L;
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.expelMember(crewDto, memberId, invalidUserId));
-
-        assertEquals(CrewErrorCode.FORBIDDEN.getMessage(), exception.getMessage());
-        then(crewPolicy).should(only())
-                .validateCrewCreator(anyLong(), anyLong());
-        then(crewMemberReader).shouldHaveNoInteractions();
-        then(meetingMemberStore).shouldHaveNoInteractions();
-        then(meetingStore).shouldHaveNoInteractions();
-        then(crewMemberStore).shouldHaveNoInteractions();
     }
 
     @DisplayName("추방하려는 크루원이 크루장일 경우, CreatorDeleteForbidden 예외를 발생시킨다.")
@@ -858,28 +480,6 @@ class CrewServiceImplTest {
 
         assertEquals(CrewErrorCode.CREATOR_DELETE_FORBIDDEN.getMessage(), exception.getMessage());
         then(crewMemberReader).shouldHaveNoInteractions();
-        then(meetingMemberStore).shouldHaveNoInteractions();
-        then(meetingStore).shouldHaveNoInteractions();
-        then(crewMemberStore).shouldHaveNoInteractions();
-    }
-
-    @DisplayName("크루원을 추방할 때, 존재하지 않는 크루원의 경우, NotCrewMember 예외를 발생시킨다.")
-    @Test
-    void when_NotExistingCrewMember_Then_NotFoundCrewMemberException2() {
-        // given
-        CrewDto crewDto = createCrewDto().build().get();
-        CrewDto temp = DomainFixtureFactory.TestCrewDto.builder().build().get();
-        Long memberId = 11L;
-        given(crewMemberReader.existsByCrewIdAndUserId(crewDto.id(), memberId))
-                .willReturn(false);
-
-        // when & then
-        Exception exception = assertThrows(BusinessException.class,
-                () -> crewService.expelMember(crewDto, memberId, crewDto.userDto().id()));
-
-        assertEquals(CrewErrorCode.NOT_CREW_MEMBER.getMessage(), exception.getMessage());
-        then(crewMemberReader).should(only())
-                .existsByCrewIdAndUserId(anyLong(), anyLong());
         then(meetingMemberStore).shouldHaveNoInteractions();
         then(meetingStore).shouldHaveNoInteractions();
         then(crewMemberStore).shouldHaveNoInteractions();
